full_PAN_OCR_API/requirements.txt
tensorflow-cpu==1.15
gunicorn
Flask==1.1.1
opencv-python-headless
pdf2image==1.13.1
Pillow==7.0.0
pytesseract==0.3.4
scikit-image==0.15.0
Werkzeug==1.0.0
numpy==1.18.1
